requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
motor==3.7.0
passlib==1.7.4
pip==22.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
python-dotenv==1.0.1
